fix: align directory CRC generation and CDC parser types

+++ crates/pithos/src/utils/conversion.rs
@@ -32,7 +32,7 @@ pub fn parse_range_input(input: &str) -> Result<Range<u64>, PithosCliError> {
     Ok(start..end)
 }
 
-pub fn parse_cdc_input(input: &str) -> Result<(u32, u32, u32), PithosCliError> {
+pub fn parse_cdc_input(input: &str) -> Result<(usize, usize, usize), PithosCliError> {
     let parts: Vec<&str> = input.split(',').collect();
     if parts.len() != 3 {
         return Err(PithosCliError::InvalidArgumentError(
@@ -40,13 +40,13 @@ pub fn parse_cdc_input(input: &str) -> Result<(u32, u32, u32), PithosCliError> {
         ));
     }
 
-    let min = parts[0].trim().parse::<u32>().map_err(|e| {
+    let min = parts[0].trim().parse::<usize>().map_err(|e| {
         PithosCliError::InvalidArgumentError(format!("Failed to parse range start: {}", e))
     })?;
-    let avg = parts[1].trim().parse::<u32>().map_err(|e| {
+    let avg = parts[1].trim().parse::<usize>().map_err(|e| {
         PithosCliError::InvalidArgumentError(format!("Failed to parse range end: {}", e))
     })?;
-    let max = parts[2].trim().parse::<u32>().map_err(|e| {
+    let max = parts[2].trim().parse::<usize>().map_err(|e| {
         PithosCliError::InvalidArgumentError(format!("Failed to parse range end: {}", e))
     })?;
